requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
protobuf
google-auth-oauthlib
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
gTTS>=2.5.0


